[Release-1.24] Replace deprecated ioutil package (#6235)
* Replace ioutil package
* check integration test null pointer
* Remove rotate retries

+++ tests/integration/integration.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"os"
 	"os/exec"
 	"os/user"
@@ -232,8 +232,10 @@ func K3sStartServer(inputArgs ...string) (*K3sServer, error) {
 
 // K3sKillServer terminates the running K3s server and its children
 func K3sKillServer(server *K3sServer) error {
-	server.log.Close()
-	os.Remove(server.log.Name())
+	if server.log != nil {
+		server.log.Close()
+		os.Remove(server.log.Name())
+	}
 	pgid, err := syscall.Getpgid(server.cmd.Process.Pid)
 	if err != nil {
 		if errors.Is(err, syscall.ESRCH) {
@@ -298,7 +300,7 @@ func K3sDumpLog(server *K3sServer) error {
 		return err
 	}
 	defer log.Close()
-	b, err := ioutil.ReadAll(log)
+	b, err := io.ReadAll(log)
 	if err != nil {
 		return err
 	}
